added subscription to seperate redux
